% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Print "FARS Read"}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string containing the name of the file
to read.  It is assumed to be a csv file (possibly compressed) 
as provided by FARS.}
}
\value{
The function returns a tibble data frame (class tbl_df) as 
   defined ' by the dplyr package.  If the file does not exist, 
   the function returns ' a NULL value.  '
}
\description{
\code{fars_read} reads data regarding fatal injuries suffered in
motor vehicle accidents in the US, as reported by the Fatality
Analysis Reporting System (FARS) of the US National Highway 
Traffic Safety Administration.
}
\examples{
df <- fars_read(make_filename(2014,dir=system.file("extdata",package="msdr3pam")))

}
\references{
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\seealso{
\code{make_filename}, \code{fars_read_years}
}
