% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Print "Make Filename"}
\usage{
make_filename(year, dir = "./inst/extdata")
}
\arguments{
\item{year}{an integer or character string specifying the 
desired (4-digit) calendar year}

\item{dir}{a character string containing the directory that contains
the FARS data}
}
\value{
The function returns the name of the file (as a character
    string) containing FARS data for the specified year
}
\description{
\code{make_filename} creates a csv filename corresponding to 
tabular data for a particular year from the Fatality Analysis 
Reporting System (FARS) of the US National Highway Traffic Safety 
Administration.  The filename follows standard FARS naming 
convention, including a "bz2"extension reflecting bzip2 compression.
}
\examples{
filename <- make_filename(2014,system.file("extdata",package="msdr3pam))
filename <- make_filename('2014',system.file("extdata",package="msdr3pam))

}
\references{
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\seealso{
\code{fars_read}
}
